# Conflicts:
#	RobotControl/RobotControlGUI/nbproject/project.properties
#	RobotControl/RobotControlGUI/src/de/naoth/rc/DialogRegistry.java
#	RobotControl/RobotControlGUI/src/de/naoth/rc/RobotControlImpl.form
#	RobotControl/RobotControlGUI/src/de/naoth/rc/RobotControlImpl.java

+++ NaoTHSoccer/Source/Cognition/Modules/VisualCortex/BallDetector/Tools/BlackSpotExtractor.cpp
@@ -100,5 +100,5 @@ int BlackSpotExtractor::calculateKeyPointsBlackBetter(const BallDetectorIntegral
     }
   }
 
-  return bestBlack.size();
+  return static_cast<int>(bestBlack.size());
 }
